+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.1.5] - 2020-04-09
+### Added
+- The `required` modifier for the `signature` block.
+### Fixed
+- Error message about blank signature not showing.
+
 ## [1.1.4] - 2020-04-08
 ### Changed
 - `DAEmpty` now accepts index assignments.